APItest.xs: fix subsignature test
THX_parse_keyword_subsignature() in APItest.xs does, amongst other
things, return a string representation of an OP_ARGCHECK op.

My commit v5.31.4-13-gf417cfa906 made the data attached to that op be a
struct rather than an array of IVs, but missed updating APItest.xs,
which this commit now amends.

This may or may not fix GH #17202 where the returned value of the
slurpy field was wrong on big-endian systems.

This is presumably due to (char)aux[2] not being the same as
aux->slurpy where one is an IV and the other is a char field.