regcomp.c: Use marker for non-user properties
A marker is created for a property that could be both a user-defined
one, and an official Unicode one.  The expansion of these is deferred
until runtime, in case the sub that implements a user-defined property
wasn't known at the time of the pattern compilation.  If at runtime,
there still isn't the proper sub, the property is looked up as
potentially an official Unicode one.

But I forgot to check the marker, which is set only if there is a
potential ambiguity.  If the marker isn't set, there is no need to try
to look it up as an official Unicode property.  I couldn't think of a
test where this fails, but at least it saves effort.