pv_uni_display: Use common fcn; \b mnemonic
This removes the (almost) duplicate code in this function to display
mnemonics for control characters that have them.  The reason the two
pieces of code aren't precisely the same is that the other function also
uses \b as a mnemonic for backspace.  Using all possible mnemonics is
desirable, so a flag is added for pv_uni_display to now use \b.  This is
now by default enabled in double-quoted strings, but not regex patterns
(as \b there means something quite different except in character classes).
B.pm is changed to expect \b.