regcomp.c: optimize away repeated NOTHING ops
Prior to this patch we optimize away chains of NOTHING ops,
but we don't mark them as OPTIMIZED which means its easy to
overlook the fact that they have been optimized away.

This marks the NOTHING ops that are actually skipped as
being OPTIMIZE so they disappear from the deparse results.