hv_func.h: tweak how we deal with 64 bit Siphash
This slightly modifies how we apply Siphash in our code
so that instead of taking the low 32 bits of the hash
we return the xor of the high and low 32 bits.

This also adds /* FALLTHROUGH */ statements to some of
the related code so that picky compilers don't get upset
as it is not a mistake.

Lastly this allows our test code to access both the
"real" 64 bit Siphash, and also the reduced 32 bit version.

Most of this is in preparation for adding tests for
some of our hash related code. See next commit.