regcomp.c: restore {} braces to DEBUG_PARSE_r multi-statements
In 15cab4d7052 the if (!SIZE_ONLY) logic was removed from regcomp.c,
but in a few places this was excessively zealous, as the braces were
removed from multiline constructs inside of DEBUG_PARSE_r macros.

EG:

    DEBUG_PARSE_r(if (!SIZE_ONLY) {
        stuff1;
        stuff2;
        stuff3;
    });

was turned into

    DEBUG_PARSE_r(
        stuff1;
        stuff2;
        stuff3;
    );

Which means that ONLY the first statement in the block was covered
by the DEBUG_PARSE_r() conditional logic. The conversion should have
been:

    DEBUG_PARSE_r({
        stuff1;
        stuff2;
        stuff3;
    });

IOW, it was necessary to preserve the {} braces in the macro call.

This silences various forms of debugging that should not be visible
in a plain

    use re 'debug';

and should only be visible with something like

    use re Debug => 'ALL';

Eg in:

    $ ./perl -Ilib -Mre=debug -le'/(foo|bar|baz)/'
    Compiling REx "(foo|bar|baz)"
    ~ tying lastbr BRANCH (9) to ender CLOSE1 (12) offset 3
    ~ tying lastbr OPEN1 (1) to ender END (14) offset 13
    Final program:
       1: OPEN1 (3)
       3:   TRIE-EXACT[bf] (12)
            <foo>
            <bar>
            <baz>
      12: CLOSE1 (14)
      14: END (0)
    stclass AHOCORASICK-EXACT[bf] minlen 3
    Freeing REx: "(foo|bar|baz)"

The "~ tying lastbr" lines are of interest pretty much only to
someone working on or maintaining the regex engine and should not
be visible to a casual user, not only because they are ugly but
also because the context to understand them is missing and they
do not help understanding how the regex operates.

(cherry picked from commit d8e1e69d20817fd4b9d94736c97ef585900451a4)