WIP - handle ${10} properly - Issue #12948
Karl asked me for some help investigating #12948, this is what I came up with.

I have not even run "make test" so likely this is broken in important ways, but
it should be enough of a thread for Karl to pull on that hopefull the whole shirt
comes apart. So to speak. :-)

What it does do is make ${10} parse the same way as $10