Porting/branch_report: add a tool to report on branch status for topic branches
This tool reviews reviews all branches, uses git-cherry to identify
branches that have already been merged and identifies them as "should
delete", and then uses a set of heuristics based on author and subject
of the commit and the number of commits in the branch to further
identify branches that are "maybe delete" (eg because the patches were
changed during merge to blead), and then puts the rest of the branches
into "keep status".

It writes four report files to disk:

* auth_branch_report.rpt

This is a breakdown by the author/email of the most recent commit on
the branch, including breakouts by "should"/"maybe"/"keep" and the
branch names.

* can_delete.rpt

This is a report of all branches that the script thinks should be
deleted.

* maybe_delete.rpt

This is a report of all branches that the script thinks are possibly
suitable for deletion.

* not_delete.rpt

This is a report of all the branches that the script is confident
that contains one or more patches that have not been merged to blead.

NOTE: This script takes some time to run and is NOT in ANY way
efficient. It could take as much as 5 or 10 minutes to run.

NOTE: The script currently assumes that the name of the "master
perl.git" remote is "origin". It operates on the branch status ON THE
REMOTE. It pays absolutely no attention to local branches, or branches
from any other remote.

BUGS: No command line options. Slow. Maybe others.