PATCH: [perl 134335], gh115,Assert fail in regmatch
This happens when the regular expression pattern, compiling under /il,
ends up with a node attempting to match one of 17 Unicode characters
that fold (/i) to multiple characters in the 0-FF range, and the match
is run when a UTF-8 locale is in effect that makes such a match legal.
The node was being marked as SIMPLE, but only nodes that match a single
character should be SIMPLE.

This ticket was originally filed as a security issue, but I don't think
it is.  If I remove the assertion, the match fails, without an
out-of-bounds access.  And we do not consider the interpreter crashing
on a DEBUGGING build a security issue.

(cherry picked from commit 2df0dd21423ddf295fe2e5d2a6665300f4c5dd94)