op.h: Move some flag bits down
This is in preparation for adding a new flag bit at the end in a future
commit.  It could have been added in the unused space that the first of
these was moved to, but the new one is less important/used, so I thought
it best to come last.  The reason to use unused space is to preserve
binary compatibility with the bits, and we don't care about that at this
point in the development cycle.