Update perlunicode base on Unicode UTS 18, regex reqs
Unicode is revising their document on what regular expression
implementations should do.  This includes retraction of a significant
part of it, which Perl did not handle (and apparently nobody else
either).  Thus we are much closer to implementing everything they say
than before.  The document is adding some new (manageable) things, which
we do not yet support.