regcomp.sym: Add new regnode type for (?[])
This new regnode is used to handle interpolated already-compiled regex
sets inside outer regex sets.

If it isn't present, it will mean that what appears to be a nested,
interpolated set really isn't.

I created a new regnode structure to hold a pointer.  This has to be
temporary as pointers can be invalidated.  I thought of just having a
regnode without a pointer as a marker, and using a parallel array to
store the data, rather than creating a whole new regnode structure for
just pointers, but parallel data structures can get out of sync, so this
seemed best.

This commit just sets up the regnode; a future commit will actually use
it.