regcomp.c: Only output warning about experimental once
If someone uses an experimental construct, prior to this commit, they
would get a warning each time they used it in the same pattern.  This
commit causes the warning to be emitted once per pattern.

I didn't add tests, because this I didn't think it important enough to
spend the time.  The consequences of this breaking in the future are
minimal, and the constructs are temporary, likely to be removed next
release.