Regenerate charclass_invlists.h
There is something wrong with our mechanism to show if this out-of
-sync, because it didn't.  And it needed regenerating.  I will have to
look to understand the reason why.  Nor did any of the tests fail.  In
part, I see from looking at the diffs that there is a rule that is no
longer used.  But it also may be that the Unicode-supplied test are
misisng things.  Obviously one can't test every code point, but just a
representative sample, so some things may fall through the cracks.