Allow wildcard pattern debugging
This fixes #17026

Patterns can now have subpatterns, called wildcards by Unicode.  Prior
to this commit debugging info was only available under DEBUGGING builds
with -Drv.  This commit adds a new keyword, WILDCARD for
'use re qw(Debug ...)' so that whatever other debugging options have
been turned on will show up when a wildcard subpattern is compiled.

The output of this may be voluminous, which is why you have to ask for
it specifically.  Or, the EXTRA option turns it on along with several
other things.