Allow wildcard pattern debugging
This fixes #17026

Patterns can have subpatterns since 5.30.  These are processed when
encountered, by suspending the main pattern compilation, compiling the
subpattern, and then matching that against the set of all legal
possibilities, which Perl knows about.

Prior to this commit, debugging info was not available for that matching
portion of the compilation, except under DEBUGGING builds, with -Drv.
This commit adds a new option to 'use re qw(Debug ...)', WILDCARD, to
enable subpattern match debugging.  Whatever other match debugging
options have been turned on will show up when a wildcard subpattern is
compiled iff WILDCARD is specified.

The output of this may be voluminous, which is why you have to ask for
it specifically.  Or, the EXTRA option turns it on, along with several
other things.