undef Tie::File objects before untie-ing arrays
This test file was added before we converted all .pm and .t files under
dist/Tie-File to 'use strict' and 'use warnings'.  So at that point the
test ran without warnings.  Once we subsequently began to run with
warnings, we failed to note two instances of "untie attempted while 1
inner references still exist" warnings.  So we should undef the
Tie::File objects before calling 'untie' on the arrays holding records
of tied files.

(Indeed, undef-ing the Tie::File objects may in itself be sufficient to
avert these warnings.  However, the pre-existing SYNOPSIS shows use of
'untie', so we'll include it in this file, whose point is to test what
is asserted in that document.)