reentr.c: revise handling of several wrapped fcns
Prior to this commit, the code in this function had generic variables,
like p0, and malloc'd them using void *, and then cast to what it
thought were the proper types.  This was failing on solaris g++, because
of the stricter casting rules in C++.

This commit actually simplifies the code presentation, by getting rid of
the generic stuff, and for each group of related functions, declares the
appropriate names for the variables, and mallocs them using the correct
type as expressed in the __DATA__ section of regen/reentr.pl.  The
casting is eliminated.

This fixes #17482.

The __DATA__ section includes what types we expect the arguments to be,
plus a list of prototypes that someone thought were possibilities for
the function for the various platforms.  I do not know how that list was
derived; perhaps looking at man pages, then finding more things from
experience.  regen/reentr.pl looks through its list of possibilities to
set up #ifdefs for those various possibilities in reentr.[ch].
Configure generates the prototypes actually existing on the platform,
and thus what compiles is what those #ifdefs evaluate to.  It is
possible that the fixed arguments we have could be wrong, but we have no
way of correcting that.  The previous implementation cast to various
types, but those turned out to not always be correct.  Perhaps the
non-re-entrant versions have more standardized parameters.  I don't
know.
come up with #ifdefThat could mean that our exceptions A more
complete solution would be to look at the prototypes generated by
Configure, and